dev_requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
loguru==0.7.2
packaging==24.2
psycopg==3.2.3
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
